refactor date-range parser
